+++ src/array/array_sync_readable.rs
@@ -5,6 +5,7 @@ use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterato
 use crate::{
     array::ArrayMetadataV2,
     array_subset::ArraySubset,
+    metadata::MetadataRetrieveVersion,
     node::NodePath,
     storage::{
         data_key, meta_key, meta_key_v2_array, meta_key_v2_attributes, ReadableStorageTraits,
@@ -28,58 +29,67 @@ use super::{
 use super::elements_to_ndarray;
 
 impl<TStorage: ?Sized + ReadableStorageTraits + 'static> Array<TStorage> {
-    /// Create an array in `storage` at `path`. The metadata is read from the store.
+    /// Open an existing array in `storage` at `path` with default [`MetadataRetrieveVersion`].
+    /// The metadata is read from the store.
     ///
     /// # Errors
     /// Returns [`ArrayCreateError`] if there is a storage error or any metadata is invalid.
+    // #[deprecated(since = "0.15.0", note = "please use `open` instead")]
     pub fn new(storage: Arc<TStorage>, path: &str) -> Result<Self, ArrayCreateError> {
-        let array_v3 = Self::new_v3(storage.clone(), path);
-        if let Err(ArrayCreateError::MissingMetadata) = array_v3 {
-            Self::new_v2(storage, path)
-        } else {
-            array_v3
-        }
+        Self::open(storage, path)
     }
 
-    /// Create a new Zarr V3 array in `storage` at `path`. The metadata is read from the store.
+    /// Open an existing array in `storage` at `path` with default [`MetadataRetrieveVersion`].
+    /// The metadata is read from the store.
     ///
     /// # Errors
     /// Returns [`ArrayCreateError`] if there is a storage error or any metadata is invalid.
-    pub fn new_v3(storage: Arc<TStorage>, path: &str) -> Result<Self, ArrayCreateError> {
-        let node_path = NodePath::new(path)?;
-        let key = meta_key(&node_path);
-        let metadata: ArrayMetadataV3 = serde_json::from_slice(
-            &storage
-                .get(&key)?
-                .ok_or(ArrayCreateError::MissingMetadata)?,
-        )
-        .map_err(|err| StorageError::InvalidMetadata(key, err.to_string()))?;
-        Self::new_with_metadata(storage, path, ArrayMetadata::V3(metadata))
+    pub fn open(storage: Arc<TStorage>, path: &str) -> Result<Self, ArrayCreateError> {
+        Self::open_opt(storage, path, &MetadataRetrieveVersion::Default)
     }
 
-    /// Create a new Zarr V2 array in `storage` at `path`. The metadata is read from the store.
+    /// Open an existing array in `storage` at `path` with non-default [`MetadataRetrieveVersion`].
+    /// The metadata is read from the store.
     ///
     /// # Errors
     /// Returns [`ArrayCreateError`] if there is a storage error or any metadata is invalid.
-    pub fn new_v2(storage: Arc<TStorage>, path: &str) -> Result<Self, ArrayCreateError> {
+    pub fn open_opt(
+        storage: Arc<TStorage>,
+        path: &str,
+        version: &MetadataRetrieveVersion,
+    ) -> Result<Self, ArrayCreateError> {
         let node_path = NodePath::new(path)?;
 
-        let metadata_key = meta_key_v2_array(&node_path);
-        let mut metadata: ArrayMetadataV2 = serde_json::from_slice(
-            &storage
-                .get(&metadata_key)?
-                .ok_or(ArrayCreateError::MissingMetadata)?,
-        )
-        .map_err(|err| StorageError::InvalidMetadata(metadata_key, err.to_string()))?;
-
-        let attributes_key = meta_key_v2_attributes(&node_path);
-        let attributes = storage.get(&attributes_key)?;
-        if let Some(attributes) = attributes {
-            metadata.attributes = serde_json::from_slice(&attributes)
-                .map_err(|err| StorageError::InvalidMetadata(attributes_key, err.to_string()))?;
+        if let MetadataRetrieveVersion::Default | MetadataRetrieveVersion::V3 = version {
+            // Try V3
+            let key_v3 = meta_key(&node_path);
+            if let Some(metadata) = storage.get(&key_v3)? {
+                let metadata: ArrayMetadataV3 = serde_json::from_slice(&metadata)
+                    .map_err(|err| StorageError::InvalidMetadata(key_v3, err.to_string()))?;
+                return Self::new_with_metadata(storage, path, ArrayMetadata::V3(metadata));
+            }
         }
 
-        Self::new_with_metadata(storage, path, ArrayMetadata::V2(metadata))
+        if let MetadataRetrieveVersion::Default | MetadataRetrieveVersion::V2 = version {
+            // Try V2
+            let key_v2 = meta_key_v2_array(&node_path);
+            if let Some(metadata) = storage.get(&key_v2)? {
+                let mut metadata: ArrayMetadataV2 = serde_json::from_slice(&metadata)
+                    .map_err(|err| StorageError::InvalidMetadata(key_v2, err.to_string()))?;
+
+                let attributes_key = meta_key_v2_attributes(&node_path);
+                let attributes = storage.get(&attributes_key)?;
+                if let Some(attributes) = attributes {
+                    metadata.attributes = serde_json::from_slice(&attributes).map_err(|err| {
+                        StorageError::InvalidMetadata(attributes_key, err.to_string())
+                    })?;
+                }
+
+                return Self::new_with_metadata(storage, path, ArrayMetadata::V2(metadata));
+            }
+        }
+
+        Err(ArrayCreateError::MissingMetadata)
     }
 
     /// Read and decode the chunk at `chunk_indices` into its bytes if it exists with default codec options.