+++ src/metadata.rs
@@ -26,10 +26,58 @@ pub use v3::{
     UnsupportedAdditionalFieldError,
 };
 
+use crate::config::global_config;
+
 /// A type alias for [`MetadataV3`].
 ///
 /// Kept for backwards compatibility with `zarrs` < 0.15.
 pub type Metadata = MetadataV3;
+
+/// The metadata version to retrieve.
+///
+/// Used with [`crate::array::Array::open_opt`], [`crate::group::Group::open_opt`].
+pub enum MetadataRetrieveVersion {
+    /// Either V3 or V2. V3 is prioritised over V2 if found.
+    Default,
+    /// V3
+    V3,
+    /// V2
+    V2,
+}
+
+/// Version options for [`Array::store_metadata`](crate::array::Array::store_metadata) and [`Group::store_metadata`](crate::group::Group::store_metadata), and their async variants.
+#[derive(Debug, Clone, Copy)]
+pub enum MetadataConvertVersion {
+    /// Write the same version as the input metadata.
+    Default,
+    /// Write Zarr V3 metadata. Zarr V2 metadata will not be automatically removed if it exists.
+    V3,
+}
+
+impl Default for MetadataConvertVersion {
+    fn default() -> Self {
+        *global_config().metadata_convert_version()
+    }
+}
+
+/// Version options for [`Array::erase_metadata`](crate::array::Array::erase_metadata) and [`Group::erase_metadata`](crate::group::Group::erase_metadata), and their async variants.
+#[derive(Debug, Clone, Copy)]
+pub enum MetadataEraseVersion {
+    /// Erase the same version as the input metadata.
+    Default,
+    /// Erase all metadata.
+    All,
+    /// Erase V3 metadata.
+    V3,
+    /// Erase V2 metadata.
+    V2,
+}
+
+impl Default for MetadataEraseVersion {
+    fn default() -> Self {
+        *global_config().metadata_erase_version()
+    }
+}
 
 #[cfg(test)]
 mod tests {